improve dependency flow related to encryption
