+++ src/encryption/index.js
@@ -1,0 +1,14 @@
+/* @flow */
+export {
+  encryptECIES,
+  decryptECIES,
+  signECDSA,
+  verifyECDSA,
+  CipherObject,
+  getHexFromBN
+} from './ec'
+
+export { 
+  encryptMnemonic,
+  decryptMnemonic
+} from './wallet'
